+++ packages/opencode/src/credentials/index.ts
@@ -2,3 +2,4 @@ export { Credentials } from "./types"
 export { CredentialStore } from "./store"
 export { CredentialsMigrate } from "./migrate"
 export { CredentialPool } from "./pool"
+export { isOAuthSecretWithRefresh, isOAuthSecret, isApiSecret } from "./guards"
